requirements.txt
# Core
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2023.3
peewee>=3.16.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Notifications
slack-sdk>=3.21.0

# Interactive Brokers API (optional — only needed if using IB sync)
# ib_insync>=0.9.86

# Testing (not needed in production)
# pytest>=7.0.0
